+++ Twilio/Rest/Voice/V1/VoicePermission/Country/HighriskSpecialPrefixInstance.php
@@ -7,7 +7,7 @@
  * /       /
  */
 
-namespace Twilio\Rest\Preview\Permissions;
+namespace Twilio\Rest\Voice\V1\VoicePermission\Country;
 
 use Twilio\Exceptions\TwilioException;
 use Twilio\InstanceResource;
@@ -16,19 +16,25 @@ use Twilio\Version;
 
 /**
  * PLEASE NOTE that this class contains preview products that are subject to change. Use them with caution. If you currently do not have developer preview access, please contact help@twilio.com.
+ * 
+ * @property string prefix
  */
-class VoicePermissionInstance extends InstanceResource {
+class HighriskSpecialPrefixInstance extends InstanceResource {
     /**
-     * Initialize the VoicePermissionInstance
+     * Initialize the HighriskSpecialPrefixInstance
      * 
      * @param \Twilio\Version $version Version that contains the resource
      * @param mixed[] $payload The response payload
-     * @return \Twilio\Rest\Preview\Permissions\VoicePermissionInstance 
+     * @param string $isoCode The ISO country code
+     * @return \Twilio\Rest\Voice\V1\VoicePermission\Country\HighriskSpecialPrefixInstance 
      */
-    public function __construct(Version $version, array $payload) {
+    public function __construct(Version $version, array $payload, $isoCode) {
         parent::__construct($version);
 
-        $this->solution = array();
+        // Marshaled Properties
+        $this->properties = array('prefix' => Values::array_get($payload, 'prefix'), );
+
+        $this->solution = array('isoCode' => $isoCode, );
     }
 
     /**
@@ -57,6 +63,6 @@ class VoicePermissionInstance extends InstanceResource {
      * @return string Machine friendly representation
      */
     public function __toString() {
-        return '[Twilio.Preview.Permissions.VoicePermissionInstance]';
+        return '[Twilio.Voice.V1.HighriskSpecialPrefixInstance]';
     }
 }