+++ Twilio/Rest/Voice/V1/VoicePermission/CountryPage.php
@@ -7,14 +7,14 @@
  * /       /
  */
 
-namespace Twilio\Rest\Preview\Permissions\VoicePermission;
+namespace Twilio\Rest\Voice\V1\VoicePermission;
 
 use Twilio\Page;
 
 /**
  * PLEASE NOTE that this class contains preview products that are subject to change. Use them with caution. If you currently do not have developer preview access, please contact help@twilio.com.
  */
-class SettingsPage extends Page {
+class CountryPage extends Page {
     public function __construct($version, $response, $solution) {
         parent::__construct($version, $response);
 
@@ -23,7 +23,7 @@ class SettingsPage extends Page {
     }
 
     public function buildInstance(array $payload) {
-        return new SettingsInstance($this->version, $payload);
+        return new CountryInstance($this->version, $payload);
     }
 
     /**
@@ -32,6 +32,6 @@ class SettingsPage extends Page {
      * @return string Machine friendly representation
      */
     public function __toString() {
-        return '[Twilio.Preview.Permissions.SettingsPage]';
+        return '[Twilio.Voice.V1.CountryPage]';
     }
 }